moved progressAlertHandler into seperate file
